Initial commit
This repository includes all the C++ code from the `src/framework` directory of the BioCro R package repository. In the future, this repository will be included in that one as a git submodule. Most references to R functions or files, with the major exceptions of `Rinternals.h` and the BioCro R package, have been removed since the code here is generally intended to be independent of the R interface.